Upgrade to arrow/parquet 55, and `object_store` to `0.12.0` and pyo3 to `0.24.0` (#15466)
* Temp pin to datafusion main

* Update cargo lock

* update pyo3

* vendor random generation

* Update error message

* Update for extraction

* Update pin

* Upgrade object_store

* fix feature

* Update file size handling

* bash for object store API changes

* few more

* Update APIs more

* update expected message

* update error messages

* Update to apache

* Update API for nicer parquet u64s

* Fix wasm build

* Remove pin

* Fix signature

+++ datafusion/core/src/datasource/file_format/arrow.rs
@@ -144,6 +144,7 @@ impl FileFormat for ArrowFormat {
         for object in objects {
             let r = store.as_ref().get(&object.location).await?;
             let schema = match r.payload {
+                #[cfg(not(target_arch = "wasm32"))]
                 GetResultPayload::File(mut file, _) => {
                     let reader = FileReader::try_new(&mut file, None)?;
                     reader.schema()
@@ -437,7 +438,7 @@ mod tests {
         let object_meta = ObjectMeta {
             location,
             last_modified: DateTime::default(),
-            size: usize::MAX,
+            size: u64::MAX,
             e_tag: None,
             version: None,
         };
@@ -480,7 +481,7 @@ mod tests {
         let object_meta = ObjectMeta {
             location,
             last_modified: DateTime::default(),
-            size: usize::MAX,
+            size: u64::MAX,
             e_tag: None,
             version: None,
         };
